requirements.txt
# Core dependencies
rados>=2.0.0  # Ceph RADOS Python bindings

# Vector embeddings
sentence-transformers>=2.2.0
torch>=2.0.0

# Vector database
chromadb>=0.4.0

# LLM and NLP (for future extensions)
transformers>=4.30.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Configuration and schemas
pyyaml>=6.0
pydantic>=2.0.0

# CLI
click>=8.1.0
rich>=13.0.0  # Beautiful terminal output
tqdm>=4.65.0  # Progress bars

# Text processing
python-magic>=0.4.27
chardet>=5.1.0

# Optional: OpenAI support (for future)
# openai>=1.0.0

# Utilities
python-dateutil>=2.8.0
